+++ lib/nadia/behaviour.ex
@@ -3,12 +3,18 @@ defmodule Nadia.Behaviour do
 
   alias Nadia.Model.{
     BotAccessSettings,
+    BotCommand,
+    BotDescription,
+    BotName,
+    BotShortDescription,
     BusinessConnection,
     Chat,
+    ChatAdministratorRights,
     ChatMember,
     Error,
     File,
     ForumTopic,
+    MenuButton,
     Message,
     MessageId,
     Poll,
@@ -38,19 +44,43 @@ defmodule Nadia.Behaviour do
   @callback delete_my_commands([{atom, any}] | map) :: :ok | {:error, Error.t()}
   @callback delete_my_commands(Client.t()) :: :ok | {:error, Error.t()}
   @callback delete_my_commands(Client.t(), [{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @callback get_my_commands() :: {:ok, [BotCommand.t()]} | {:error, Error.t()}
+  @callback get_my_commands([{atom, any}] | map) ::
+              {:ok, [BotCommand.t()]} | {:error, Error.t()}
+  @callback get_my_commands(Client.t()) :: {:ok, [BotCommand.t()]} | {:error, Error.t()}
+  @callback get_my_commands(Client.t(), [{atom, any}] | map) ::
+              {:ok, [BotCommand.t()]} | {:error, Error.t()}
   @callback set_my_name() :: :ok | {:error, Error.t()}
   @callback set_my_name([{atom, any}] | map) :: :ok | {:error, Error.t()}
   @callback set_my_name(Client.t()) :: :ok | {:error, Error.t()}
   @callback set_my_name(Client.t(), [{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @callback get_my_name() :: {:ok, BotName.t()} | {:error, Error.t()}
+  @callback get_my_name([{atom, any}] | map) :: {:ok, BotName.t()} | {:error, Error.t()}
+  @callback get_my_name(Client.t()) :: {:ok, BotName.t()} | {:error, Error.t()}
+  @callback get_my_name(Client.t(), [{atom, any}] | map) ::
+              {:ok, BotName.t()} | {:error, Error.t()}
   @callback set_my_description() :: :ok | {:error, Error.t()}
   @callback set_my_description([{atom, any}] | map) :: :ok | {:error, Error.t()}
   @callback set_my_description(Client.t()) :: :ok | {:error, Error.t()}
   @callback set_my_description(Client.t(), [{atom, any}] | map) :: :ok | {:error, Error.t()}
+  @callback get_my_description() :: {:ok, BotDescription.t()} | {:error, Error.t()}
+  @callback get_my_description([{atom, any}] | map) ::
+              {:ok, BotDescription.t()} | {:error, Error.t()}
+  @callback get_my_description(Client.t()) :: {:ok, BotDescription.t()} | {:error, Error.t()}
+  @callback get_my_description(Client.t(), [{atom, any}] | map) ::
+              {:ok, BotDescription.t()} | {:error, Error.t()}
   @callback set_my_short_description() :: :ok | {:error, Error.t()}
   @callback set_my_short_description([{atom, any}] | map) :: :ok | {:error, Error.t()}
   @callback set_my_short_description(Client.t()) :: :ok | {:error, Error.t()}
   @callback set_my_short_description(Client.t(), [{atom, any}] | map) ::
               :ok | {:error, Error.t()}
+  @callback get_my_short_description() :: {:ok, BotShortDescription.t()} | {:error, Error.t()}
+  @callback get_my_short_description([{atom, any}] | map) ::
+              {:ok, BotShortDescription.t()} | {:error, Error.t()}
+  @callback get_my_short_description(Client.t()) ::
+              {:ok, BotShortDescription.t()} | {:error, Error.t()}
+  @callback get_my_short_description(Client.t(), [{atom, any}] | map) ::
+              {:ok, BotShortDescription.t()} | {:error, Error.t()}
   @callback remove_my_profile_photo() :: :ok | {:error, Error.t()}
   @callback remove_my_profile_photo([{atom, any}] | map) :: :ok | {:error, Error.t()}
   @callback remove_my_profile_photo(Client.t()) :: :ok | {:error, Error.t()}
@@ -61,12 +91,26 @@ defmodule Nadia.Behaviour do
   @callback set_chat_menu_button(Client.t()) :: :ok | {:error, Error.t()}
   @callback set_chat_menu_button(Client.t(), [{atom, any}] | map) ::
               :ok | {:error, Error.t()}
+  @callback get_chat_menu_button() :: {:ok, MenuButton.t()} | {:error, Error.t()}
+  @callback get_chat_menu_button([{atom, any}] | map) ::
+              {:ok, MenuButton.t()} | {:error, Error.t()}
+  @callback get_chat_menu_button(Client.t()) :: {:ok, MenuButton.t()} | {:error, Error.t()}
+  @callback get_chat_menu_button(Client.t(), [{atom, any}] | map) ::
+              {:ok, MenuButton.t()} | {:error, Error.t()}
   @callback set_my_default_administrator_rights() :: :ok | {:error, Error.t()}
   @callback set_my_default_administrator_rights([{atom, any}] | map) ::
               :ok | {:error, Error.t()}
   @callback set_my_default_administrator_rights(Client.t()) :: :ok | {:error, Error.t()}
   @callback set_my_default_administrator_rights(Client.t(), [{atom, any}] | map) ::
               :ok | {:error, Error.t()}
+  @callback get_my_default_administrator_rights() ::
+              {:ok, ChatAdministratorRights.t()} | {:error, Error.t()}
+  @callback get_my_default_administrator_rights([{atom, any}] | map) ::
+              {:ok, ChatAdministratorRights.t()} | {:error, Error.t()}
+  @callback get_my_default_administrator_rights(Client.t()) ::
+              {:ok, ChatAdministratorRights.t()} | {:error, Error.t()}
+  @callback get_my_default_administrator_rights(Client.t(), [{atom, any}] | map) ::
+              {:ok, ChatAdministratorRights.t()} | {:error, Error.t()}
   @callback set_user_emoji_status(integer) :: :ok | {:error, Error.t()}
   @callback set_user_emoji_status(integer, [{atom, any}] | map) ::
               :ok | {:error, Error.t()}
@@ -556,24 +600,42 @@ defmodule Nadia.Behaviour do
                       delete_my_commands: 0,
                       delete_my_commands: 1,
                       delete_my_commands: 2,
+                      get_my_commands: 0,
+                      get_my_commands: 1,
+                      get_my_commands: 2,
                       set_my_name: 0,
                       set_my_name: 1,
                       set_my_name: 2,
+                      get_my_name: 0,
+                      get_my_name: 1,
+                      get_my_name: 2,
                       set_my_description: 0,
                       set_my_description: 1,
                       set_my_description: 2,
+                      get_my_description: 0,
+                      get_my_description: 1,
+                      get_my_description: 2,
                       set_my_short_description: 0,
                       set_my_short_description: 1,
                       set_my_short_description: 2,
+                      get_my_short_description: 0,
+                      get_my_short_description: 1,
+                      get_my_short_description: 2,
                       remove_my_profile_photo: 0,
                       remove_my_profile_photo: 1,
                       remove_my_profile_photo: 2,
                       set_chat_menu_button: 0,
                       set_chat_menu_button: 1,
                       set_chat_menu_button: 2,
+                      get_chat_menu_button: 0,
+                      get_chat_menu_button: 1,
+                      get_chat_menu_button: 2,
                       set_my_default_administrator_rights: 0,
                       set_my_default_administrator_rights: 1,
                       set_my_default_administrator_rights: 2,
+                      get_my_default_administrator_rights: 0,
+                      get_my_default_administrator_rights: 1,
+                      get_my_default_administrator_rights: 2,
                       set_user_emoji_status: 1,
                       set_user_emoji_status: 2,
                       set_user_emoji_status: 3,